+++ include/lss/generator/generator.hpp
@@ -8,7 +8,7 @@ public:
   Generator();
   std::shared_ptr<Location> getLocation(LocationSpec);
   std::shared_ptr<Location> getRandomLocation(std::shared_ptr<Player>,
-                                              int depth = -1);
+                                              int depth = -1, std::shared_ptr<Cell> enter = nullptr);
 };
 
 #endif // __GENERATOR_H_